requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# dev / test only
fakeredis[lua]>=2.21.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
